FIXME: Re-structure eav-seq to be able to be schema aware.
This is in preparation for adding :cardinality/many functionality.

FIXME NOTE: We really don't need or want duplicate schema and attrs data. The
schema is formatted in the datomic/datascript way but that's not useful for
lookup so we convert to the attrs format. Since we can easily convert attrs
back to schema, let's just keep one version. We can also just call it 'schema
instead of 'attrs which should make more sense, I think.

Co-authored-by: Joel Martin <joel.martin@viasat.com>