[rules] FIX comparison of different value types would throw
When trying to identify colliding 'insertables and retractables, 'cmp is not
safe for values of differing types. This could happen in either the :e or :v
positions (:a positions are always keywords). This new approach simply compares
hash values as we don't care about actual ordering, only stable ordering and
comparison.

Refactored out complex in-line code as 'bagwise-mutual-diff which yields both
the difference of A-B and B-A while preserving multiplicity, which set
difference would not. This is because we need to know not just that a
particular :e :a :v appears in both 'insertables and 'retractables but what is
the net. i.e. they shouldn't cancel each other out if there are more on one
side or the other.

Co-Authored-By: Joel Martin <joel.martin@viasat.com>