[store,rules] New 'transact with proper checks
In order to do proper checking, we need the context of the full transaction.
This includes an initial rules/transact and store/transact*. Checking does not
happen for rules/retract (and kin) or rules/upsert (and kin). There also isn't
a rules/transact! yet.

This will support a more datomic/datascript like interaction.

Co-Authored-By: Joel Martin <joel.martin@viasat.com>