Add 2x3 test cases. Continue on exceptions.
Iterate across all test cases that contain two transactions with each
transaction containing 1, 2, or three Aeavs (Aeav is action + eav). Each Aeav is
8 possible combinations of:
  - A: :db/add or :db/retract
  - e: -1 or -2 (tempids)
  - a: always :schema/name
  - v: "X" or "O"

Additionally, accumulate the states of the databases/sessions after each
transaction and compare the final sequence of states for equality. One of the
states can be an exception rather than a list of entities. For generative
testing that greatly increases the scope of the tests because previous we were
typically stopping early in the test case right after the first transaction.
Now we apply the entire test case and compare all the intermediate and final
results.

As part of implementing this test scenario, refactor the code in run-txs into a
reduce on run-tx where the databases/sessions are passed in as a context
object. run-tx itself calls out to database specific entity gathering functions.

Includes some interesting test cases as comments at the bottom for comparing
clara-factbase and Datomic behavior.

Co-Authored-By: Joel Martin <joel.martin@viasat.com>