Avoid tx's with different behavior from Datomic. Add ordered datomic DB.
Add a function that checks a transaction (and current state of the database)
for scenarios where Datomic behavior is know to differ. This function is used
to trigger an exception across all database/sessions which in effect ignores
the transaction (while still capturing the fact that it was ignored in the results).

This increases test coverage during generative testing because we will explore
larger test cases that include the ignore transactions (otherwise we mark this
as a failure and begin shriking before exploring further).

Also, add a top level DB type :datomic-ordered that does the sorting of
retracts before adds within single transactions. The :datomic DB is switched
back to use the original order of the transaction. This allows us to detect
differences in behavior from datomic due to ordering of Aeavs within a tx.

This includes a refactor of run-tx to split out each database/session
application function (run-*-tx). Also, adds run-generative-cycles function
which repeatedly runs a generative test function.

Captures some additional interesting cases in a comment at the bottom.

Co-Authored-By: Joel Martin <joel.martin@viasat.com>